[cleanup] factorize tutorials in examples
* collects duplicate code across tutorials in helpers.py
