+++ examples/tutorial_pipe.py
@@ -1,16 +1,14 @@
 import torch
-import torch.nn as nn
-import torch.nn.functional as F
 import torch.optim as optim
 
 import fairscale
+from helpers import getModel, getData, getLossFun
 
 DEVICE = "cuda" if torch.cuda.is_available() else "cpu"
 
-model = nn.Sequential(torch.nn.Linear(10, 10), torch.nn.ReLU(), torch.nn.Linear(10, 5))
-target = torch.randint(0, 2, size=(20, 1)).squeeze()
-data = torch.randn(20, 10)
-loss_fn = F.nll_loss
+model = getModel()
+data, target = getData()
+loss_fn = getLossFun()
 
 model = fairscale.nn.Pipe(model, balance=[2, 1])
