+++ examples/tutorial_pipe_multiprocess.py
@@ -3,21 +3,14 @@ import os
 import torch
 import torch.distributed as dist
 import torch.multiprocessing as mp
-import torch.nn as nn
-import torch.nn.functional as F
 import torch.optim as optim
 
 import fairscale
 from fairscale.nn.model_parallel import initialize_model_parallel
+from helpers import dist_init, getModel, getData, getLossFun
 
 
 DEVICE = "cuda" if torch.cuda.is_available() else "cpu"
-
-
-def dist_init(rank, world_size):
-    backend = "nccl" if torch.cuda.is_available() else "gloo"
-    print(f"Using backend: {backend}")
-    dist.init_process_group(backend=backend, rank=rank, world_size=world_size)
 
 
 def run(rank, world_size):
@@ -28,10 +21,9 @@ def run(rank, world_size):
     dist.rpc.init_rpc(f"worker{rank}", rank=rank, world_size=world_size)
     initialize_model_parallel(1, world_size)
 
-    model = nn.Sequential(torch.nn.Linear(10, 10), torch.nn.ReLU(), torch.nn.Linear(10, 5))
-    target = torch.randint(0, 2, size=(20, 1)).squeeze()
-    data = torch.randn(20, 10)
-    loss_fn = F.nll_loss
+    model = getModel()
+    data, target = getData()
+    loss_fn = getLossFun()
 
     device = torch.device("cuda", rank) if DEVICE == "cuda" else torch.device("cpu")
